рефакторинг
убирает дублирования функций в view компонентах
